refactor: `PresetManager` モジュールを統合
