+++ build_util/make_docs.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from voicevox_engine.app.application import generate_app
 from voicevox_engine.dev.core.mock import MockCoreWrapper
 from voicevox_engine.dev.tts_engine.mock import MockTTSEngine
-from voicevox_engine.preset.PresetManager import PresetManager
+from voicevox_engine.preset.Preset import PresetManager
 from voicevox_engine.setting.SettingLoader import USER_SETTING_PATH, SettingHandler
 from voicevox_engine.tts_pipeline.tts_engine import CoreAdapter
 from voicevox_engine.user_dict.user_dict import UserDictionary
